+++ Utils/D19cSSA2Event.h
@@ -53,7 +53,7 @@ class D19cSSA2Event : public Event
         }
     }
 
-    void fillChipDataContainer(ChipDataContainer* chipContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup, uint16_t hybridId) override;
+    void fillChipDataContainer(ChipDataContainer* chipContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup, uint8_t hybridId) override;
 
   private:
     uint8_t                   fNSSA2 = 1;
