+++ Utils/Event.h
@@ -408,7 +408,7 @@ class Event
 
     virtual std::vector<Cluster> getClusters(uint8_t pHybridId, uint8_t pCbcId) const { return {}; }
     virtual void                 fillDataContainer(BoardDataContainer* boardContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup);
-    virtual void                 fillChipDataContainer(ChipDataContainer* boardContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup, uint16_t hybridId) = 0;
+    virtual void                 fillChipDataContainer(ChipDataContainer* boardContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup, uint8_t hybridId) = 0;
 
     // split stream of data
     template <std::size_t N>
